DT-2754 Refactor; split UserResource so account creation linking and locking are in their own resource file
