+++ src/main/kotlin/uk/gov/justice/digital/hmpps/nomisuserrolesapi/resource/UserManagementResource.kt
@@ -1,0 +1,146 @@
+package uk.gov.justice.digital.hmpps.nomisuserrolesapi.resource
+
+import io.swagger.v3.oas.annotations.Operation
+import io.swagger.v3.oas.annotations.media.Content
+import io.swagger.v3.oas.annotations.media.Schema
+import io.swagger.v3.oas.annotations.responses.ApiResponse
+import io.swagger.v3.oas.annotations.security.SecurityRequirement
+import org.springframework.http.HttpStatus
+import org.springframework.http.MediaType
+import org.springframework.security.access.prepost.PreAuthorize
+import org.springframework.validation.annotation.Validated
+import org.springframework.web.bind.annotation.PathVariable
+import org.springframework.web.bind.annotation.PutMapping
+import org.springframework.web.bind.annotation.RequestBody
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.ResponseStatus
+import org.springframework.web.bind.annotation.RestController
+import uk.gov.justice.digital.hmpps.nomisuserrolesapi.config.AuthenticationFacade
+import uk.gov.justice.digital.hmpps.nomisuserrolesapi.config.ErrorResponse
+import uk.gov.justice.digital.hmpps.nomisuserrolesapi.service.UserService
+import javax.validation.Valid
+import javax.validation.constraints.Pattern
+import javax.validation.constraints.Size
+
+@RestController
+@Validated
+@RequestMapping("/users", produces = [MediaType.APPLICATION_JSON_VALUE])
+class UserManagementResource(
+  private val userService: UserService,
+  private val authenticationFacade: AuthenticationFacade,
+) {
+
+  @PreAuthorize("hasRole('ROLE_MANAGE_NOMIS_USER_ACCOUNT')")
+  @PutMapping("/{username}/lock-user")
+  @ResponseStatus(HttpStatus.OK)
+  @Operation(
+    summary = "Lock user account",
+    description = "Locks the user account. Requires role ROLE_MANAGE_NOMIS_USER_ACCOUNT",
+    security = [SecurityRequirement(name = "MANAGE_NOMIS_USER_ACCOUNT")],
+    responses = [
+      ApiResponse(
+        responseCode = "200",
+        description = "User account locked"
+      ),
+      ApiResponse(
+        responseCode = "400",
+        description = "Incorrect request to lock user",
+        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
+      ),
+      ApiResponse(
+        responseCode = "401",
+        description = "Unauthorized to access this endpoint",
+        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
+      ),
+      ApiResponse(
+        responseCode = "403",
+        description = "Incorrect permissions to lock a user",
+        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
+      )
+    ]
+  )
+  fun lockUser(
+    @Schema(description = "Username", example = "testuser1", required = true)
+    @PathVariable @Size(max = 30, min = 1, message = "username must be between 1 and 30") username: String
+  ) {
+    userService.lockUser(username)
+  }
+
+  @PreAuthorize("hasRole('ROLE_MANAGE_NOMIS_USER_ACCOUNT')")
+  @PutMapping("/{username}/unlock-user")
+  @ResponseStatus(HttpStatus.OK)
+  @Operation(
+    summary = "Unlock user account",
+    description = "Unlocks the user account. Requires role ROLE_MANAGE_NOMIS_USER_ACCOUNT",
+    security = [SecurityRequirement(name = "MANAGE_NOMIS_USER_ACCOUNT")],
+    responses = [
+      ApiResponse(
+        responseCode = "200",
+        description = "User account unlocked"
+      ),
+      ApiResponse(
+        responseCode = "400",
+        description = "Incorrect request to unlock user",
+        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
+      ),
+      ApiResponse(
+        responseCode = "401",
+        description = "Unauthorized to access this endpoint",
+        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
+      ),
+      ApiResponse(
+        responseCode = "403",
+        description = "Incorrect permissions to unlock a user",
+        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
+      )
+    ]
+  )
+  fun unlockUser(
+    @Schema(description = "Username", example = "testuser1", required = true)
+    @PathVariable @Size(max = 30, min = 1, message = "username must be between 1 and 30") username: String
+  ) {
+    userService.unlockUser(username)
+  }
+
+  @PreAuthorize("hasRole('ROLE_MANAGE_NOMIS_USER_ACCOUNT')")
+  @PutMapping("/{username}/change-password")
+  @ResponseStatus(HttpStatus.OK)
+  @Operation(
+    summary = "Change password of user account",
+    description = "Change password of user account. Requires role ROLE_MANAGE_NOMIS_USER_ACCOUNT",
+    security = [SecurityRequirement(name = "MANAGE_NOMIS_USER_ACCOUNT")],
+    responses = [
+      ApiResponse(
+        responseCode = "200",
+        description = "User account password changed"
+      ),
+      ApiResponse(
+        responseCode = "400",
+        description = "Incorrect request to change password of user",
+        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
+      ),
+      ApiResponse(
+        responseCode = "401",
+        description = "Unauthorized to access this endpoint",
+        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
+      ),
+      ApiResponse(
+        responseCode = "403",
+        description = "Incorrect permissions to change the password a user",
+        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
+      )
+    ]
+  )
+  fun changePassword(
+    @Schema(description = "Username", example = "testuser1", required = true)
+    @PathVariable @Size(max = 30, min = 1, message = "username must be between 1 and 30") username: String,
+    @Schema(description = "Password", example = "HeLl0W0R1D", required = true)
+    @Pattern(
+      regexp = "^[A-Za-z0-9]{14,30}$",
+      message = "Password must consist of alphanumeric characters only and a minimum of 14 chars, and max 30 chars"
+    )
+    @RequestBody @Valid password: String,
+  ) {
+    userService.changePassword(username, password)
+  }
+}
