+++ src/main/kotlin/uk/gov/justice/digital/hmpps/nomisuserrolesapi/resource/UserResource.kt
@@ -10,34 +10,23 @@ import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.data.domain.Sort
 import org.springframework.data.web.PageableDefault
-import org.springframework.http.HttpStatus
 import org.springframework.http.MediaType
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.validation.annotation.Validated
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.PutMapping
-import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
-import org.springframework.web.bind.annotation.ResponseStatus
 import org.springframework.web.bind.annotation.RestController
 import uk.gov.justice.digital.hmpps.nomisuserrolesapi.config.AuthenticationFacade
 import uk.gov.justice.digital.hmpps.nomisuserrolesapi.config.ErrorResponse
-import uk.gov.justice.digital.hmpps.nomisuserrolesapi.data.CreateAdminUserRequest
-import uk.gov.justice.digital.hmpps.nomisuserrolesapi.data.CreateGeneralUserRequest
-import uk.gov.justice.digital.hmpps.nomisuserrolesapi.data.CreateLinkedAdminUserRequest
-import uk.gov.justice.digital.hmpps.nomisuserrolesapi.data.CreateLinkedGeneralUserRequest
 import uk.gov.justice.digital.hmpps.nomisuserrolesapi.data.StaffDetail
 import uk.gov.justice.digital.hmpps.nomisuserrolesapi.data.UserDetail
 import uk.gov.justice.digital.hmpps.nomisuserrolesapi.data.UserStatus
 import uk.gov.justice.digital.hmpps.nomisuserrolesapi.data.UserSummary
 import uk.gov.justice.digital.hmpps.nomisuserrolesapi.data.filter.UserFilter
 import uk.gov.justice.digital.hmpps.nomisuserrolesapi.service.UserService
-import javax.validation.Valid
-import javax.validation.constraints.Pattern
 import javax.validation.constraints.Size
 
 @RestController
@@ -47,275 +36,6 @@ class UserResource(
   private val userService: UserService,
   private val authenticationFacade: AuthenticationFacade,
 ) {
-  @PreAuthorize("hasRole('ROLE_CREATE_USER')")
-  @PostMapping("/general-account")
-  @ResponseStatus(HttpStatus.CREATED)
-  @Operation(
-    summary = "Create general user account",
-    description = "Creates general user account, oracle schema and staff user information. Requires role ROLE_CREATE_USER",
-    security = [SecurityRequirement(name = "CREATE_USER")],
-    requestBody = io.swagger.v3.oas.annotations.parameters.RequestBody(
-      content = [
-        Content(
-          mediaType = "application/json",
-          schema = Schema(implementation = CreateGeneralUserRequest::class)
-        )
-      ]
-    ),
-    responses = [
-      ApiResponse(
-        responseCode = "201",
-        description = "General user information returned"
-      ),
-      ApiResponse(
-        responseCode = "400",
-        description = "Incorrect request to create user information",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Incorrect permissions to create a user",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      )
-    ]
-  )
-  fun createGeneralUser(
-    @RequestBody @Valid createUserRequest: CreateGeneralUserRequest
-  ): UserDetail {
-    val user = userService.createGeneralUser(createUserRequest)
-    return userService.findByUsername(username = user.username)
-  }
-
-  @PreAuthorize("hasRole('ROLE_CREATE_USER')")
-  @PostMapping("/local-admin-account")
-  @ResponseStatus(HttpStatus.CREATED)
-  @Operation(
-    summary = "Create local admin user account",
-    description = "Creates local admin user account, oracle schema and staff user information. Requires role ROLE_CREATE_USER",
-    security = [SecurityRequirement(name = "CREATE_USER")],
-    requestBody = io.swagger.v3.oas.annotations.parameters.RequestBody(
-      content = [
-        Content(
-          mediaType = "application/json",
-          schema = Schema(implementation = CreateGeneralUserRequest::class)
-        )
-      ]
-    ),
-    responses = [
-      ApiResponse(
-        responseCode = "201",
-        description = "Local Admin user information returned"
-      ),
-      ApiResponse(
-        responseCode = "400",
-        description = "Incorrect request to create user information",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Incorrect permissions to create a user",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      )
-    ]
-  )
-  fun createLocalAdminUser(
-    @RequestBody @Valid createLocalAdminUserRequest: CreateGeneralUserRequest
-  ): UserDetail {
-    val user = userService.createLocalAdminUser(createLocalAdminUserRequest)
-    return userService.findByUsername(username = user.username)
-  }
-
-  @PreAuthorize("hasRole('ROLE_CREATE_USER')")
-  @PostMapping("/admin-account")
-  @ResponseStatus(HttpStatus.CREATED)
-  @Operation(
-    summary = "Create admin user account",
-    description = "Creates admin user account, oracle schema and staff user information. Requires role ROLE_CREATE_USER",
-    security = [SecurityRequirement(name = "CREATE_USER")],
-    requestBody = io.swagger.v3.oas.annotations.parameters.RequestBody(
-      content = [
-        Content(
-          mediaType = "application/json",
-          schema = Schema(implementation = CreateAdminUserRequest::class)
-        )
-      ]
-    ),
-    responses = [
-      ApiResponse(
-        responseCode = "201",
-        description = "Admin user account information returned"
-      ),
-      ApiResponse(
-        responseCode = "400",
-        description = "Incorrect request to create user information",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Incorrect permissions to create a user",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      )
-    ]
-  )
-  fun createAdminUser(
-    @RequestBody @Valid createUserRequest: CreateAdminUserRequest
-  ): UserDetail {
-    val user = userService.createAdminUser(createUserRequest)
-    return userService.findByUsername(username = user.username)
-  }
-
-  @PreAuthorize("hasRole('ROLE_CREATE_USER')")
-  @PostMapping("/link-general-account/{linkedUsername}")
-  @ResponseStatus(HttpStatus.CREATED)
-  @Operation(
-    summary = "Link a general user account to an existing admin account.",
-    description = "Can only be linked to an admin account. Can only be linked to an account that doesn't already have one general account. Requires role ROLE_CREATE_USER",
-    security = [SecurityRequirement(name = "CREATE_USER")],
-    requestBody = io.swagger.v3.oas.annotations.parameters.RequestBody(
-      content = [
-        Content(
-          mediaType = "application/json",
-          schema = Schema(implementation = CreateLinkedGeneralUserRequest::class)
-        )
-      ]
-    ),
-    responses = [
-      ApiResponse(
-        responseCode = "201",
-        description = "Staff account information returned"
-      ),
-      ApiResponse(
-        responseCode = "400",
-        description = "Incorrect request to link general account",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Incorrect permissions to link a general account",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      )
-    ]
-  )
-  fun linkGeneralAccount(
-    @Schema(description = "Attach account to an existing admin user account", example = "testuser2", required = true)
-    @PathVariable @Size(max = 30, min = 1, message = "Username must be between 1 and 30") linkedUsername: String,
-    @RequestBody @Valid linkedGeneralUserRequest: CreateLinkedGeneralUserRequest
-  ): StaffDetail {
-    val linkedUser = userService.linkGeneralAccount(linkedUsername, linkedGeneralUserRequest)
-    return userService.findByStaffId(linkedUser.staffId)
-  }
-
-  @PreAuthorize("hasRole('ROLE_CREATE_USER')")
-  @PostMapping("/link-local-admin-account/{linkedUsername}")
-  @ResponseStatus(HttpStatus.CREATED)
-  @Operation(
-    summary = "Link a local admin user account to an existing general account.",
-    description = "Can only be linked to an general account. Can only be linked to an account that doesn't already have one admin account. Requires role ROLE_CREATE_USER",
-    security = [SecurityRequirement(name = "CREATE_USER")],
-    requestBody = io.swagger.v3.oas.annotations.parameters.RequestBody(
-      content = [
-        Content(
-          mediaType = "application/json",
-          schema = Schema(implementation = CreateLinkedGeneralUserRequest::class)
-        )
-      ]
-    ),
-    responses = [
-      ApiResponse(
-        responseCode = "201",
-        description = "Staff local admin account information returned"
-      ),
-      ApiResponse(
-        responseCode = "400",
-        description = "Incorrect request to link local admin account",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Incorrect permissions to link a local admin account",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      )
-    ]
-  )
-  fun linkLocalAdminAccount(
-    @Schema(description = "Attach account to an existing general user account", example = "testuser2", required = true)
-    @PathVariable @Size(max = 30, min = 1, message = "Username must be between 1 and 30") linkedUsername: String,
-    @RequestBody @Valid linkedLocalAdminUserRequest: CreateLinkedGeneralUserRequest
-  ): StaffDetail {
-    val linkedUser = userService.linkLocalAdminAccount(linkedUsername, linkedLocalAdminUserRequest)
-    return userService.findByStaffId(linkedUser.staffId)
-  }
-
-  @PreAuthorize("hasRole('ROLE_CREATE_USER')")
-  @PostMapping("/link-admin-account/{linkedUsername}")
-  @ResponseStatus(HttpStatus.CREATED)
-  @Operation(
-    summary = "Link an admin account to an existing general account.",
-    description = "Can only be linked to an general account. Can only be linked to an account that doesn't already have one Admin account. Requires role ROLE_CREATE_USER",
-    security = [SecurityRequirement(name = "CREATE_USER")],
-    requestBody = io.swagger.v3.oas.annotations.parameters.RequestBody(
-      content = [
-        Content(
-          mediaType = "application/json",
-          schema = Schema(implementation = CreateLinkedAdminUserRequest::class)
-        )
-      ]
-    ),
-    responses = [
-      ApiResponse(
-        responseCode = "201",
-        description = "Staff account information returned"
-      ),
-      ApiResponse(
-        responseCode = "400",
-        description = "Incorrect request to link admin account",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Incorrect permissions to link an admin account",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      )
-    ]
-  )
-  fun linkAdminAccount(
-    @Schema(description = "Attach account to an existing general account", example = "testuser2", required = true)
-    @PathVariable @Size(max = 30, min = 1, message = "Username must be between 1 and 30") linkedUsername: String,
-    @RequestBody @Valid linkedUserRequest: CreateLinkedAdminUserRequest
-  ): StaffDetail {
-    val linkedUser = userService.linkAdminAccount(linkedUsername, linkedUserRequest)
-    return userService.findByStaffId(linkedUser.staffId)
-  }
 
   @PreAuthorize("hasRole('ROLE_CREATE_USER')")
   @DeleteMapping("/{username}")
@@ -477,117 +197,6 @@ class UserResource(
       roleCodes = accessRoles ?: listOf(),
     ),
   )
-
-  @PreAuthorize("hasRole('ROLE_MANAGE_NOMIS_USER_ACCOUNT')")
-  @PutMapping("/{username}/lock-user")
-  @ResponseStatus(HttpStatus.OK)
-  @Operation(
-    summary = "Lock user account",
-    description = "Locks the user account. Requires role ROLE_MANAGE_NOMIS_USER_ACCOUNT",
-    security = [SecurityRequirement(name = "MANAGE_NOMIS_USER_ACCOUNT")],
-    responses = [
-      ApiResponse(
-        responseCode = "200",
-        description = "User account locked"
-      ),
-      ApiResponse(
-        responseCode = "400",
-        description = "Incorrect request to lock user",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Incorrect permissions to lock a user",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      )
-    ]
-  )
-  fun lockUser(
-    @Schema(description = "Username", example = "testuser1", required = true)
-    @PathVariable @Size(max = 30, min = 1, message = "username must be between 1 and 30") username: String
-  ) {
-    userService.lockUser(username)
-  }
-
-  @PreAuthorize("hasRole('ROLE_MANAGE_NOMIS_USER_ACCOUNT')")
-  @PutMapping("/{username}/unlock-user")
-  @ResponseStatus(HttpStatus.OK)
-  @Operation(
-    summary = "Unlock user account",
-    description = "Unlocks the user account. Requires role ROLE_MANAGE_NOMIS_USER_ACCOUNT",
-    security = [SecurityRequirement(name = "MANAGE_NOMIS_USER_ACCOUNT")],
-    responses = [
-      ApiResponse(
-        responseCode = "200",
-        description = "User account unlocked"
-      ),
-      ApiResponse(
-        responseCode = "400",
-        description = "Incorrect request to unlock user",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Incorrect permissions to unlock a user",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      )
-    ]
-  )
-  fun unlockUser(
-    @Schema(description = "Username", example = "testuser1", required = true)
-    @PathVariable @Size(max = 30, min = 1, message = "username must be between 1 and 30") username: String
-  ) {
-    userService.unlockUser(username)
-  }
-
-  @PreAuthorize("hasRole('ROLE_MANAGE_NOMIS_USER_ACCOUNT')")
-  @PutMapping("/{username}/change-password")
-  @ResponseStatus(HttpStatus.OK)
-  @Operation(
-    summary = "Change password of user account",
-    description = "Change password of user account. Requires role ROLE_MANAGE_NOMIS_USER_ACCOUNT",
-    security = [SecurityRequirement(name = "MANAGE_NOMIS_USER_ACCOUNT")],
-    responses = [
-      ApiResponse(
-        responseCode = "200",
-        description = "User account password changed"
-      ),
-      ApiResponse(
-        responseCode = "400",
-        description = "Incorrect request to change password of user",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Incorrect permissions to change the password a user",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))]
-      )
-    ]
-  )
-  fun changePassword(
-    @Schema(description = "Username", example = "testuser1", required = true)
-    @PathVariable @Size(max = 30, min = 1, message = "username must be between 1 and 30") username: String,
-    @Schema(description = "Password", example = "HeLl0W0R1D", required = true)
-    @Pattern(regexp = "^[A-Za-z0-9]{14,30}$", message = "Password must consist of alphanumeric characters only and a minimum of 14 chars, and max 30 chars")
-    @RequestBody @Valid password: String,
-  ) {
-    userService.changePassword(username, password)
-  }
 
   fun localAdministratorUsernameWhenNotCentralAdministrator(): String? =
     if (AuthenticationFacade.hasRoles("ROLE_MAINTAIN_ACCESS_ROLES_ADMIN")) null else authenticationFacade.currentUsername
