+++ src/test/kotlin/uk/gov/justice/digital/hmpps/nomisuserrolesapi/resource/UserResourceIntTest.kt
@@ -1,6 +1,5 @@
 package uk.gov.justice.digital.hmpps.nomisuserrolesapi.resource
 
-import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.DisplayName
@@ -8,14 +7,9 @@ import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.web.reactive.function.BodyInserters
-import uk.gov.justice.digital.hmpps.nomisuserrolesapi.data.CreateAdminUserRequest
 import uk.gov.justice.digital.hmpps.nomisuserrolesapi.data.CreateGeneralUserRequest
-import uk.gov.justice.digital.hmpps.nomisuserrolesapi.data.CreateLinkedAdminUserRequest
-import uk.gov.justice.digital.hmpps.nomisuserrolesapi.data.CreateLinkedGeneralUserRequest
-import uk.gov.justice.digital.hmpps.nomisuserrolesapi.data.StaffDetail
 import uk.gov.justice.digital.hmpps.nomisuserrolesapi.helper.DataBuilder
 import uk.gov.justice.digital.hmpps.nomisuserrolesapi.integration.IntegrationTestBase
-import uk.gov.justice.digital.hmpps.nomisuserrolesapi.jpa.UsageType
 
 class UserResourceIntTest : IntegrationTestBase() {
   @Autowired
@@ -602,6 +596,7 @@ class UserResourceIntTest : IntegrationTestBase() {
           .jsonPath("totalPages").isEqualTo(11)
           .jsonPath("size").isEqualTo(10)
       }
+
       @Test
       internal fun `can request a different page size`() {
         webTestClient.get().uri {
@@ -621,6 +616,7 @@ class UserResourceIntTest : IntegrationTestBase() {
           .jsonPath("size").isEqualTo(20)
           .jsonPath("$.content[0].username").isEqualTo("aaabella.moulin")
       }
+
       @Test
       internal fun `can request a different page`() {
         webTestClient.get().uri {
@@ -642,309 +638,6 @@ class UserResourceIntTest : IntegrationTestBase() {
           .jsonPath("$.content[0].username").isEqualTo("another.user60")
           .jsonPath("$.content[1].username").isEqualTo("another.user61")
       }
-    }
-  }
-
-  @Nested
-  @DisplayName("POST /users")
-  inner class MaintainUsers {
-
-    @Test
-    fun `a database user cannot be created without correct role`() {
-
-      webTestClient.post().uri("/users/general-account")
-        .headers(setAuthorisation(roles = listOf("ROLE_DUMMY")))
-        .body(
-          BodyInserters.fromValue(
-            CreateGeneralUserRequest(
-              username = "testuser2",
-              firstName = "Test",
-              lastName = "User",
-              defaultCaseloadId = "BXI",
-              email = "test@test.com"
-            )
-          )
-        )
-        .exchange()
-        .expectStatus().isForbidden
-    }
-    @Test
-    fun `a local admin database user can be created`() {
-
-      webTestClient.post().uri("/users/local-admin-account")
-        .headers(setAuthorisation(roles = listOf("ROLE_CREATE_USER")))
-        .body(
-          BodyInserters.fromValue(
-            CreateGeneralUserRequest(
-              username = "laauser1",
-              firstName = "laa",
-              lastName = "User",
-              email = "laa@test.com",
-              defaultCaseloadId = "PVI"
-            )
-          )
-        )
-        .exchange()
-        .expectStatus().isCreated
-        .expectBody().json(
-          """
-          {
-          "username": "LAAUSER1",
-          "firstName": "LAA",
-          "lastName": "USER",
-          "activeCaseloadId" : "PVI",
-          "active": true,
-          "accountStatus": "EXPIRED",
-          "primaryEmail": "laa@test.com",
-          "accountType": "ADMIN"
-          }
-          """
-        )
-    }
-
-    @Test
-    fun `an local admin user can be linked to  general database user can local admin can search for general user`() {
-
-      webTestClient.post().uri("/users/local-admin-account")
-        .headers(setAuthorisation(roles = listOf("ROLE_CREATE_USER")))
-        .body(
-          BodyInserters.fromValue(
-            CreateGeneralUserRequest(
-              username = "laa_user1",
-              firstName = "laa",
-              lastName = "User",
-              email = "laa@test.com",
-              defaultCaseloadId = "WWI"
-            )
-          )
-        )
-        .exchange()
-        .expectStatus().isCreated
-        .expectBody().json(
-          """
-          {
-          "username": "LAA_USER1",
-          "firstName": "LAA",
-          "lastName": "USER",
-          "activeCaseloadId" : "WWI",
-          "active": true,
-          "accountStatus": "EXPIRED",
-          "primaryEmail": "laa@test.com",
-          "accountType": "ADMIN"
-          }
-          """
-        )
-
-      webTestClient.post().uri("/users/link-general-account/LAA_USER1")
-        .headers(setAuthorisation(roles = listOf("ROLE_CREATE_USER")))
-        .body(
-          BodyInserters.fromValue(
-            CreateLinkedGeneralUserRequest(
-              username = "generaluser1",
-              defaultCaseloadId = "WWI"
-            )
-          )
-        )
-        .exchange()
-        .expectStatus().isCreated
-        .expectBody()
-        .jsonPath("generalAccount.username").isEqualTo("GENERALUSER1")
-        .jsonPath("generalAccount.activeCaseload.id").isEqualTo("WWI")
-        .jsonPath("generalAccount.active").isEqualTo("true")
-        .jsonPath("generalAccount.accountType").isEqualTo("GENERAL")
-        .jsonPath("adminAccount.username").isEqualTo("LAA_USER1")
-        .jsonPath("adminAccount.activeCaseload.id").isEqualTo("WWI")
-        .jsonPath("adminAccount.active").isEqualTo("true")
-        .jsonPath("adminAccount.accountType").isEqualTo("ADMIN")
-
-      webTestClient.get().uri { it.path("/users/").queryParam("nameFilter", "generaluser1").build() }
-        .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_ACCESS_ROLES"), user = "LAA_USER1"))
-        .exchange()
-        .expectStatus().isOk
-        .expectBody()
-        .jsonPath("$.numberOfElements").isEqualTo(1)
-        .jsonPath("$.content[?(@.username == '%s')]", "GENERALUSER1").exists()
-    }
-
-    @Test
-    fun `a admin database user can be created`() {
-
-      webTestClient.post().uri("/users/admin-account")
-        .headers(setAuthorisation(roles = listOf("ROLE_CREATE_USER")))
-        .body(
-          BodyInserters.fromValue(
-            CreateAdminUserRequest(
-              username = "testuser2",
-              firstName = "Test",
-              lastName = "User",
-              email = "test@test.com"
-            )
-          )
-        )
-        .exchange()
-        .expectStatus().isCreated
-        .expectBody().json(
-          """
-          {
-          "username": "TESTUSER2",
-          "firstName": "TEST",
-          "lastName": "USER",
-          "activeCaseloadId" : "CADM_I",
-          "active": true,
-          "accountStatus": "EXPIRED",
-          "primaryEmail": "test@test.com",
-          "accountType": "ADMIN"
-          }
-          """
-        )
-
-      webTestClient.get().uri("/users/TESTUSER2")
-        .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_ACCESS_ROLES_ADMIN")))
-        .exchange()
-        .expectStatus().isOk
-        .expectBody()
-        .jsonPath("username").isEqualTo("TESTUSER2")
-        .jsonPath("firstName").isEqualTo("TEST")
-        .jsonPath("lastName").isEqualTo("USER")
-        .jsonPath("staffId").exists()
-    }
-
-    @Test
-    fun `a user can be link to an existing account`() {
-
-      webTestClient.post().uri("/users/admin-account")
-        .headers(setAuthorisation(roles = listOf("ROLE_CREATE_USER")))
-        .body(
-          BodyInserters.fromValue(
-            CreateAdminUserRequest(
-              username = "testuser4",
-              firstName = "Test",
-              lastName = "User",
-              email = "test@test.com"
-            )
-          )
-        )
-        .exchange()
-        .expectStatus().isCreated
-
-      val staffDetail = webTestClient.post().uri("/users/link-general-account/TESTUSER4")
-        .headers(setAuthorisation(roles = listOf("ROLE_CREATE_USER")))
-        .body(
-          BodyInserters.fromValue(
-            CreateLinkedGeneralUserRequest(
-              username = "testuser5",
-              defaultCaseloadId = "BXI"
-            )
-          )
-        )
-        .exchange()
-        .expectStatus().isCreated
-        .expectBody(StaffDetail::class.java)
-        .returnResult().responseBody!!
-
-      assertThat(staffDetail.firstName).isEqualTo("TEST")
-      assertThat(staffDetail.lastName).isEqualTo("USER")
-      assertThat(staffDetail.primaryEmail).isEqualTo("test@test.com")
-      assertThat(staffDetail.status).isEqualTo("ACTIVE")
-
-      assertThat(staffDetail.generalAccount).isNotNull
-      assertThat(staffDetail.generalAccount?.username).isEqualTo("TESTUSER5")
-      assertThat(staffDetail.generalAccount?.activeCaseload?.id).isEqualTo("BXI")
-      assertThat(staffDetail.generalAccount?.active).isEqualTo(true)
-      assertThat(staffDetail.generalAccount?.accountType).isEqualTo(UsageType.GENERAL)
-
-      assertThat(staffDetail.adminAccount).isNotNull
-      assertThat(staffDetail.adminAccount?.username).isEqualTo("TESTUSER4")
-      assertThat(staffDetail.adminAccount?.activeCaseload?.id).isEqualTo("CADM_I")
-      assertThat(staffDetail.adminAccount?.active).isEqualTo(true)
-      assertThat(staffDetail.adminAccount?.accountType).isEqualTo(UsageType.ADMIN)
-
-      webTestClient.get().uri("/users/staff/${staffDetail.staffId}")
-        .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_ACCESS_ROLES_ADMIN")))
-        .exchange()
-        .expectStatus().isOk
-        .expectBody()
-        .jsonPath("staffId").exists()
-        .jsonPath("firstName").isEqualTo("TEST")
-        .jsonPath("lastName").isEqualTo("USER")
-        .jsonPath("primaryEmail").isEqualTo("test@test.com")
-        .jsonPath("status").isEqualTo("ACTIVE")
-        .jsonPath("generalAccount.username").isEqualTo("TESTUSER5")
-        .jsonPath("generalAccount.activeCaseload.id").isEqualTo("BXI")
-        .jsonPath("generalAccount.active").isEqualTo("true")
-        .jsonPath("generalAccount.accountType").isEqualTo("GENERAL")
-        .jsonPath("adminAccount.username").isEqualTo("TESTUSER4")
-        .jsonPath("adminAccount.activeCaseload.id").isEqualTo("CADM_I")
-        .jsonPath("adminAccount.active").isEqualTo("true")
-        .jsonPath("adminAccount.accountType").isEqualTo("ADMIN")
-    }
-
-    @Test
-    fun `a user cannot be link to an existing account of same ADMIN type`() {
-
-      webTestClient.post().uri("/users/admin-account")
-        .headers(setAuthorisation(roles = listOf("ROLE_CREATE_USER")))
-        .body(
-          BodyInserters.fromValue(
-            CreateAdminUserRequest(
-              username = "testuser6",
-              firstName = "Test",
-              lastName = "User",
-              email = "test@test.com"
-            )
-          )
-        )
-        .exchange()
-        .expectStatus().isCreated
-
-      webTestClient.post().uri("/users/link-admin-account/TESTUSER6")
-        .headers(setAuthorisation(roles = listOf("ROLE_CREATE_USER")))
-        .body(
-          BodyInserters.fromValue(
-            CreateLinkedAdminUserRequest(
-              username = "testuser5"
-            )
-          )
-        )
-        .exchange()
-        .expectStatus().is4xxClientError
-        .expectBody()
-        .jsonPath("userMessage").isEqualTo("User already exists: Admin user already exists for this staff member")
-    }
-
-    @Test
-    fun `a user cannot be link to an existing account of same GENERAL type`() {
-
-      webTestClient.post().uri("/users/general-account")
-        .headers(setAuthorisation(roles = listOf("ROLE_CREATE_USER")))
-        .body(
-          BodyInserters.fromValue(
-            CreateGeneralUserRequest(
-              username = "testuser7",
-              firstName = "Test",
-              lastName = "User",
-              defaultCaseloadId = "BXI",
-              email = "test@test.com"
-            )
-          )
-        )
-        .exchange()
-        .expectStatus().isCreated
-
-      webTestClient.post().uri("/users/link-general-account/TESTUSER7")
-        .headers(setAuthorisation(roles = listOf("ROLE_CREATE_USER")))
-        .body(
-          BodyInserters.fromValue(
-            CreateLinkedGeneralUserRequest(
-              username = "testuser8",
-              defaultCaseloadId = "BXI"
-            )
-          )
-        )
-        .exchange()
-        .expectStatus().is4xxClientError
-        .expectBody()
-        .jsonPath("userMessage").isEqualTo("User already exists: General user already exists for this staff member")
     }
   }
 
@@ -995,138 +688,6 @@ class UserResourceIntTest : IntegrationTestBase() {
         .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_ACCESS_ROLES_ADMIN")))
         .exchange()
         .expectStatus().isNotFound
-    }
-  }
-
-  @Nested
-  @DisplayName("PUT /users/{username}/lock-user")
-  inner class LockUnlockUsers {
-
-    @BeforeEach
-    internal fun createUsers() {
-      webTestClient.post().uri("/users/general-account")
-        .headers(setAuthorisation(roles = listOf("ROLE_CREATE_USER")))
-        .body(
-          BodyInserters.fromValue(
-            CreateGeneralUserRequest(
-              username = "LOCKING_USER1",
-              firstName = "Locking",
-              lastName = "User",
-              email = "test@test.com",
-              defaultCaseloadId = "PVI"
-            )
-          )
-        )
-        .exchange()
-        .expectStatus().isCreated
-    }
-
-    @AfterEach
-    internal fun deleteUsers() {
-      webTestClient.delete().uri("/users/LOCKING_USER1")
-        .headers(setAuthorisation(roles = listOf("ROLE_CREATE_USER")))
-        .exchange()
-        .expectStatus().isOk
-    }
-
-    @Test
-    fun `can't lock a db user that doesn't exist`() {
-      webTestClient.put().uri("/users/LOCKING_USER2/lock-user")
-        .headers(setAuthorisation(roles = listOf("ROLE_MANAGE_NOMIS_USER_ACCOUNT")))
-        .exchange()
-        .expectStatus().isNotFound
-    }
-
-    @Test
-    fun `can't lock a db user without correct role`() {
-      webTestClient.put().uri("/users/LOCKING_USER1/lock-user")
-        .headers(setAuthorisation(roles = listOf("ROLE_DUMMY")))
-        .exchange()
-        .expectStatus().isForbidden
-    }
-
-    @Test
-    fun `can lock a db user that does exist`() {
-      webTestClient.put().uri("/users/LOCKING_USER1/lock-user")
-        .headers(setAuthorisation(roles = listOf("ROLE_MANAGE_NOMIS_USER_ACCOUNT")))
-        .exchange()
-        .expectStatus().isOk
-
-      webTestClient.get().uri("/users/LOCKING_USER1")
-        .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_ACCESS_ROLES_ADMIN")))
-        .exchange()
-        .expectStatus().isOk
-        .expectBody()
-        .jsonPath("accountStatus").isEqualTo("LOCKED")
-    }
-
-    @Test
-    fun `can't unlock a db user that doesn't exist`() {
-      webTestClient.put().uri("/users/LOCKING_USER2/unlock-user")
-        .headers(setAuthorisation(roles = listOf("ROLE_MANAGE_NOMIS_USER_ACCOUNT")))
-        .exchange()
-        .expectStatus().isNotFound
-    }
-
-    @Test
-    fun `can't unlock a db user without correct role`() {
-      webTestClient.put().uri("/users/LOCKING_USER1/unlock-user")
-        .headers(setAuthorisation(roles = listOf("ROLE_DUMMY")))
-        .exchange()
-        .expectStatus().isForbidden
-    }
-
-    @Test
-    fun `can unlock a db user that does exist`() {
-      webTestClient.put().uri("/users/LOCKING_USER1/unlock-user")
-        .headers(setAuthorisation(roles = listOf("ROLE_MANAGE_NOMIS_USER_ACCOUNT")))
-        .exchange()
-        .expectStatus().isOk
-
-      webTestClient.get().uri("/users/LOCKING_USER1")
-        .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_ACCESS_ROLES_ADMIN")))
-        .exchange()
-        .expectStatus().isOk
-        .expectBody()
-        .jsonPath("accountStatus").isEqualTo("OPEN")
-    }
-
-    @Test
-    fun `can't change password of a db user that doesn't exist`() {
-      webTestClient.put().uri("/users/LOCKING_USER2/change-password")
-        .headers(setAuthorisation(roles = listOf("ROLE_MANAGE_NOMIS_USER_ACCOUNT")))
-        .body(BodyInserters.fromValue("He110W0R1D5555"))
-        .exchange()
-        .expectStatus().isNotFound
-    }
-
-    @Test
-    fun `can't change password of a db user without correct role`() {
-      webTestClient.put().uri("/users/LOCKING_USER1/change-password")
-        .headers(setAuthorisation(roles = listOf("ROLE_DUMMY")))
-        .body(BodyInserters.fromValue("He110W0R1D5555"))
-        .exchange()
-        .expectStatus().isForbidden
-    }
-
-    @Test
-    fun `can change password of a  db user that does exist`() {
-      webTestClient.put().uri("/users/LOCKING_USER1/change-password")
-        .headers(setAuthorisation(roles = listOf("ROLE_MANAGE_NOMIS_USER_ACCOUNT")))
-        .body(BodyInserters.fromValue("He110W0R1D5555"))
-        .exchange()
-        .expectStatus().isOk
-    }
-
-    @Test
-    fun `can't change password of a db user for an invalid password`() {
-      webTestClient.put().uri("/users/LOCKING_USER1/change-password")
-        .headers(setAuthorisation(roles = listOf("ROLE_MANAGE_NOMIS_USER_ACCOUNT")))
-        .body(BodyInserters.fromValue("HELLO^%$"))
-        .exchange()
-        .expectStatus().is4xxClientError
-        .expectBody()
-        .jsonPath("userMessage").isEqualTo("Validation failure: changePassword.password: Password must consist of alphanumeric characters only and a minimum of 14 chars, and max 30 chars")
     }
   }
 }
